back_end/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.0.0 